Organizing files into folders. Changing all paths in all scripts to reflect new organization.
