+++ figures/mlp_learning_curve.png
