+++ figures/svm_learning_curve.png
