+++ scripts/ADABOOST_Testing.py
@@ -1,41 +1,37 @@
 import pandas as pd
 import numpy as np
-from sklearn.preprocessing import OneHotEncoder, StandardScaler
-from sklearn.compose import ColumnTransformer
+import matplotlib.pyplot as plt
 from sklearn.model_selection import train_test_split, RepeatedStratifiedKFold, GridSearchCV, learning_curve
 from sklearn.ensemble import AdaBoostClassifier
-from sklearn.metrics import accuracy_score, classification_report, confusion_matrix, roc_auc_score, matthews_corrcoef, \
+from sklearn.metrics import accuracy_score, confusion_matrix, classification_report, roc_auc_score, matthews_corrcoef, \
     roc_curve
-import matplotlib.pyplot as plt
+from sklearn.tree import DecisionTreeClassifier
 
-# Load the original dataset
-filename = '/Users/lubainakothari/Documents/BINF6100MLproject/rynazal_abundance_metadata_tranformation.csv'
+# Update matplotlib settings for plotting
+plt.rcParams.update({
+    'font.size': 12,      # global font size
+    'font.family': 'sans-serif',
+    'axes.labelsize': 12,  # font size of the axes labels
+    'axes.titlesize': 16,  # font size of the axes title
+    'xtick.labelsize': 12,  # font size of the tick labels
+    'ytick.labelsize': 12   # font size of the tick labels
+})
+
+# Define the location of the dataset
+filename = '../filtered_data/rynazal_filtered_abundance.csv'
+
+# Load the dataset; header is the first row
 df = pd.read_csv(filename, header=0)
 
-# Specify categorical and numeric features
-categorical_features = ['gender', 'country']
-numeric_features = df.drop(['Sample ID', 'CRC', 'gender', 'country'], axis=1).columns.tolist()
-
-# Create preprocessors for numeric and categorical data
-numeric_transformer = StandardScaler()
-categorical_transformer = OneHotEncoder(drop='first')
-
-# Create a column transformer to apply the transformations
-preprocessor = ColumnTransformer(
-    transformers=[
-        ('num', numeric_transformer, numeric_features),
-        ('cat', categorical_transformer, categorical_features)
-    ])
-
-# Apply transformations
-X_processed = preprocessor.fit_transform(df.drop(['Sample ID', 'CRC'], axis=1))
+# Exclude 'Sample ID' and 'CRC' from the features, 'CRC' is the target variable
+X = df.drop(['Sample ID', 'CRC'], axis=1).values
 y = df['CRC'].values
 
 # Split into train and test sets
-X_train, X_test, y_train, y_test = train_test_split(X_processed, y, train_size=0.80, random_state=25, stratify=y)
+X_train, X_test, y_train, y_test = train_test_split(X, y, train_size=0.80, random_state=25, stratify=y)
 
-# Define the AdaBoost model with default base estimator
-model = AdaBoostClassifier(algorithm='SAMME', random_state=25)
+# Define the AdaBoost model with a Decision Tree as base estimator
+model = AdaBoostClassifier(estimator=DecisionTreeClassifier(max_depth=1), algorithm='SAMME', random_state=25)
 
 # Define the grid of values to search
 grid = {
@@ -80,34 +76,8 @@ plt.plot(fpr, tpr, label='AdaBoost (area = %0.2f)' % roc_auc)
 plt.plot([0, 1], [0, 1], 'k--')
 plt.xlabel('False Positive Rate')
 plt.ylabel('True Positive Rate')
-plt.title('ROC Curve')
+plt.title('ROC Curve for AdaBoost')
 plt.legend(loc="lower right")
-plt.show()
-
-# Get feature names from the column transformer
-feature_names = numeric_features + \
-    list(preprocessor.named_transformers_['cat'].get_feature_names_out(categorical_features))
-
-# Evaluate the model on the test set
-y_pred = best_model.predict(X_test)
-y_pred_proba = best_model.predict_proba(X_test)[:, 1]  # for ROC AUC
-
-# Calculate feature importances and display the top 10
-feature_importances = best_model.feature_importances_
-importances_df = pd.DataFrame({
-    'feature': feature_names,
-    'importance': feature_importances
-})
-importances_df = importances_df.sort_values(by='importance', ascending=False)
-
-# Plotting the top 10 features
-plt.figure(figsize=(10, 6))
-importances_df[:10].plot(kind='barh', x='feature', y='importance', legend=False, color='skyblue')
-plt.gca().invert_yaxis()
-plt.title('Top 10 Feature Importances in AdaBoost Model')
-plt.xlabel('Relative Importance')
-plt.ylabel('Features')
-plt.tight_layout()
 plt.show()
 
 # Obtain the learning curve data
@@ -138,7 +108,8 @@ fig, axes = plt.subplots(1, 3, figsize=(20, 5))
 axes[0].set_title("Learning Curve (AdaBoost)")
 axes[0].set_xlabel("Training examples")
 axes[0].set_ylabel("Score")
-axes[0].grid()
+axes[0].set_ylim(0.5, 1.01)  # Set the limits for the Y-axis
+axes[0].grid(True)
 axes[0].fill_between(train_sizes, train_scores_mean - train_scores_std,
                      train_scores_mean + train_scores_std, alpha=0.1, color="r")
 axes[0].fill_between(train_sizes, test_scores_mean - test_scores_std,
@@ -148,21 +119,46 @@ axes[0].plot(train_sizes, test_scores_mean, 'o-', color="g", label="Cross-valida
 axes[0].legend(loc="best")
 
 # Scalability plot
-axes[1].grid()
+axes[1].grid(True)
 axes[1].plot(train_sizes, fit_times_mean, 'o-')
 axes[1].fill_between(train_sizes, fit_times_mean - fit_times_std,
                      fit_times_mean + fit_times_std, alpha=0.1)
 axes[1].set_xlabel("Training examples")
 axes[1].set_ylabel("fit_times")
 axes[1].set_title("Scalability of the model")
+axes[1].set_ylim(0)  # Start y-axis at 0 for the scalability plot
 
 # Performance plot
-axes[2].grid()
+axes[2].grid(True)
 axes[2].plot(fit_times_mean, test_scores_mean, 'o-')
 axes[2].fill_between(fit_times_mean, test_scores_mean - test_scores_std,
                      test_scores_mean + test_scores_std, alpha=0.1)
 axes[2].set_xlabel("fit_times")
 axes[2].set_ylabel("Score")
 axes[2].set_title("Performance of the model")
+
+plt.show()
+
+# Feature Importance
+feature_importances = best_model.feature_importances_
+# Create a pandas series with feature importances and labels, then sort it
+importances = pd.Series(feature_importances, index=df.columns[1:-1])
+sorted_features = importances.sort_values(ascending=False)
+# Select the top 10 features
+top_importances = sorted_features[:10]
+
+# Create the plot with the specified aesthetics
+plt.figure(figsize=(10, 6))
+top_importances.plot(kind='barh', color='skyblue')
+
+# Invert y-axis to have the highest importance at the top
+plt.gca().invert_yaxis()
+
+plt.title('Top 10 Feature Importances in AdaBoost Model')
+plt.xlabel('Relative Importance')
+plt.ylabel('Features')
+
+# Tight layout to improve the spacing between subplots
+plt.tight_layout()
 
 plt.show()
