+++ scripts/randomforest_crc_metadata.py
@@ -1,9 +1,12 @@
-# Evaluate a random forest algorithm for classification
-# Reference: https://machinelearningmastery.com/random-forest-ensemble-in-python/
+from sklearn.preprocessing import OneHotEncoder, StandardScaler
+from sklearn.compose import ColumnTransformer
+from sklearn.model_selection import train_test_split, RepeatedStratifiedKFold, GridSearchCV, learning_curve
+from sklearn.metrics import accuracy_score, classification_report, confusion_matrix, roc_auc_score
+import matplotlib.pyplot as plt
 import pandas as pd
 import numpy as np
 from sklearn.model_selection import train_test_split, cross_val_score, RepeatedStratifiedKFold
-from sklearn.metrics import accuracy_score, confusion_matrix, classification_report, f1_score, matthews_corrcoef, precision_score, recall_score, roc_curve, roc_auc_score, auc
+from sklearn.metrics import accuracy_score, confusion_matrix, classification_report, f1_score, matthews_corrcoef, precision_score, recall_score, roc_curve, roc_auc_score, auc, RocCurveDisplay
 from sklearn.model_selection import learning_curve
 from sklearn.ensemble import RandomForestClassifier
 import matplotlib.pyplot as plt
@@ -11,6 +14,7 @@ from sklearn.model_selection import RandomizedSearchCV
 from pprint import pprint
 import shap
 import time
+from sklearn.pipeline import Pipeline
 
 # Define functions
 def plot_learning_curve(estimator, title, X, y, axes=None, ylim=None, cv=None,
@@ -69,63 +73,64 @@ def plot_learning_curve(estimator, title, X, y, axes=None, ylim=None, cv=None,
 
     return plt
 
+def get_feature_names_out(column_transformer):
+    """Get feature names from all transformers."""
+    new_feature_names = []
 
-"""
-Train and evaluate initial random forest model
-"""
+    # Loop over all transformers in the Column Transformer
+    for name, estimator, features in column_transformer.transformers_:
+        if name == 'remainder':
+            # The remainder are those columns not specified in transformers; they are passed as is
+            new_feature_names.extend(column_transformer.feature_names_in_[features])
+        elif hasattr(estimator, 'get_feature_names_out'):
+            # If the estimator has 'get_feature_names_out', use it
+            names = estimator.get_feature_names_out(features)
+            new_feature_names.extend(names)
+        else:
+            # Otherwise, use the provided feature names directly
+            new_feature_names.extend(features)
 
-# Define dataset
-filename = 'rynazal_filtered_abundance.csv'
+    return new_feature_names
 
-# Load the dataset; header is first row
+
+
+# Load the original dataset
+filename = '../filtered_data/rynazal_abundance_metadata.csv'
 df = pd.read_csv(filename, header=0)
 
-# Extract input (X) and output (y) variables
-data = df.values
-X = data[:, 1:-1]
-y = data[:, -1].astype(int)
+# Specify categorical and numeric features
+categorical_features = ['gender', 'country']
+numeric_features = df.drop(['Sample ID', 'CRC', 'gender', 'country'], axis=1).columns.tolist()
 
-# Split data into train and test sets
-X_train, X_test, y_train, y_test = train_test_split(X, y, train_size=0.80, random_state=25, stratify=y)
+# Create preprocessors for numeric and categorical data
+numeric_transformer = StandardScaler()
+categorical_transformer = OneHotEncoder(drop='first')
+
+# Create a column transformer to apply the transformations
+preprocessor = ColumnTransformer(
+    transformers=[
+        ('num', numeric_transformer, numeric_features),
+        ('cat', categorical_transformer, categorical_features)
+    ],
+    remainder='passthrough'  # This will pass through any other columns not listed explicitly
+)
+
+# Apply transformations
+X_processed = preprocessor.fit_transform(df.drop(['Sample ID', 'CRC'], axis=1))
+y = df['CRC'].values
+
+# Split into train and test sets
+X_train, X_test, y_train, y_test = train_test_split(X_processed, y, train_size=0.80, random_state=25, stratify=y)
+
 
 # Define the model
 model = RandomForestClassifier(random_state=25)
 
 # Train and evaluate the model with cross validation
 cv = RepeatedStratifiedKFold(n_splits=10, n_repeats=3, random_state=25)
-n_scores = cross_val_score(model, X_train, y_train, scoring='accuracy', cv=cv, n_jobs=-1, error_score='raise')
-
-# Report performance based on mean and standard deviation of accuracy scores from cross validation
-print('Accuracy: %.3f (%.3f)' % (np.mean(n_scores), np.std(n_scores)))
-
-# Based on best model from cross validation, train the model again on entire training dataset
-model.fit(X_train, y_train)
-
-# Predict output values based on the trained model with input values from the test set
-y_pred = model.predict(X_test)
-y_probs = model.predict_proba(X_test)[:,1]
-
-
-#predictions_custom = (y_probs >= 0.6).astype(int)
-
-
-# Determine accuracy scores of predictions of the test set
-accuracy = accuracy_score(y_test, y_pred)
-print(f'Accuracy: {accuracy:.3f}')
-
-# Generate a confusion matrix
-conf_matrix = confusion_matrix(y_test, y_pred)
-print('Confusion Matrix:\n', conf_matrix)
-
-# Display a classification report
-print(classification_report(y_test, y_pred))
-
-# Determine AUC
-roc_auc = roc_auc_score(y_test, y_pred)
-print("ROC AUC for og model:", roc_auc)
 
 '''
-Hyperparameter Tuning
+#Hyperparameter Tuning
 '''
 # Show default parameters
 print('Parameters currently in use:\n')
@@ -147,18 +152,14 @@ random_grid = {'n_estimators': n_estimators,
                'min_samples_leaf': min_samples_leaf,
                'bootstrap': bootstrap}
 
-pprint(random_grid)
+print(random_grid)
 
-start_time = time.time()
 
-model_random = RandomizedSearchCV(estimator=model, param_distributions=random_grid, n_iter=100, cv=5, verbose=1,
+model_random = RandomizedSearchCV(estimator=model, param_distributions=random_grid, n_iter=100, cv=cv, verbose=1,
                                   random_state=25, n_jobs=-1)
 
 model_random.fit(X_train, y_train)
 
-end_time = time.time()
-duration = end_time - start_time
-print(f"Hyperparameter tuning took {duration:.2f} seconds.")
 
 best_model = model_random.best_estimator_
 print("Best params of tuned model", model_random.best_params_)
@@ -193,10 +194,6 @@ print("Specificity:", specificity_tuned)
 
 # Plot ROC curve
 
-fpr_tuned, tpr_tuned, thresholds = roc_curve(y_test, y_probs_tuned)
-roc_auc_tuned = auc(fpr_tuned, tpr_tuned)
-print("ROC AUC for tuned model:", roc_auc_tuned)
-
 plt.figure(figsize=(8, 6))  # Optionally specify the figure size
 plt.plot(fpr_tuned, tpr_tuned, lw=2, label=f'AUC = {roc_auc_tuned:.2f}')  # Adjust color to blue
 plt.plot([0, 1], [0, 1], 'r--', label='Chance (area = 0.50)')  # Red dashed line for the chance line
@@ -204,45 +201,49 @@ plt.title('ROC Curve for Random Forest')
 plt.xlabel('False Positive Rate')
 plt.ylabel('True Positive Rate')
 plt.legend(loc="lower right")
-plt.savefig('roc_curve_tuned_ab.png')  # Save the figure
+plt.savefig('../figures/roc_curve_tuned_metadata.png')  # Save the figure
 plt.close()
 
 # Plot Feature Importance
 importances = best_model.feature_importances_
-feature_importances = pd.Series(importances, index=df.columns[1:-1])
+# Getting new feature names after transformations
+new_feature_names = get_feature_names_out(preprocessor)
+
+# Assuming `best_model` has been defined and trained
+feature_importances = pd.Series(best_model.feature_importances_, index=new_feature_names)
+
 sorted_features = feature_importances.sort_values(ascending=False)
 top_ten_features = sorted_features[:10]
 
-plt.figure(figsize=(8, 6))  # Set the figure size as needed
-
-# Sort the features in ascending order for display
-sorted_features = top_ten_features.sort_values(ascending=True)
-sorted_features.plot(kind='barh', color='skyblue')
-
+plt.figure(figsize=(12, 6))  # Increase figure size
+top_ten_features.plot(kind='bar', color='skyblue')
 plt.title('Top 10 Feature Importances Using Mean Decrease Gini')
-plt.ylabel('Features')
-plt.xlabel('Mean Decrease Gini')
-plt.tight_layout()
-plt.savefig('feature_importance_ab.png', bbox_inches='tight')
+plt.xlabel('Features')
+plt.ylabel('Mean Decrease Gini')
+plt.xticks(rotation=45)  # Rotate x-ticks to prevent overlapping
+plt.tight_layout()  # Automatically adjust subplot params to give specified padding
+plt.savefig('../figures/feature_importance.png', bbox_inches='tight')
 plt.close()
-
 
 # Plot learning curve
 
 # Assuming the previous code blocks have been executed and `best_model` is available
 fig, axes = plt.subplots(1, 3, figsize=(15, 5))
 
+# Define cross-validation strategy
+cv = RepeatedStratifiedKFold(n_splits=10, n_repeats=3, random_state=25)
+
 # Call the function
 learning_curve_fig = plot_learning_curve(estimator=best_model, title='Learning Curves (Random Forest)',
                     X=X_train, y=y_train, axes=axes, ylim=(0.7, 1.01),
                     cv=cv, n_jobs=-1)
 
-learning_curve_fig.savefig('learning_curve.png')
+learning_curve_fig.savefig('../figures/learning_curve.png')
 plt.close()
 
 # SHAP analysis
 
-explainer = shap.TreeExplainer(model)
+explainer = shap.TreeExplainer(best_model)
 shap_values = explainer.shap_values(X_test)
 
 class_index = 1  # or 0, depending on which class you are interested in
@@ -263,5 +264,5 @@ top_feature_names = df.columns[1:-1].to_numpy()[sorted_feature_indices]
 
 # Generate a summary plot for the top 10 features for the selected class
 shap.summary_plot(top_shap_values, X_test[:, sorted_feature_indices], feature_names=top_feature_names.tolist())
-plt.savefig('shap_summary_plot_top_10_class_' + str(class_index) + '.png')
+plt.savefig('../figures/shap_summary_plot_top_10_class_' + str(class_index) + '.png')
 plt.close()
